Make sure intermediate build operations are emitted
This commit simplifies the JUnit Platform test execution listener, so that
it also generates test events for "intermediate" tests. The previous implementation
was mostly copied from the legacy JUnit listener, which didn't account for
nested test structures.

There is one drawback of the approach used in this PR, which is that
because there wasn't such a concept before, we had to make a decision on
how to expose the nested tests. We're using "test classes", because it
matches closely the concept of a "test suite" which is already in use.
But it means that for those intermediate operations, the method name isn't
properly available: one has to rely on the display name, which will
actually resolve to the method name by default.

See #5975